fixed: inner class removed
